Original support document for analyze
A terrorist attack is an attack made by extremists not in line with international morality, and not limited to civilians and civilian facilities. Since 1990s, the numbers of terrorism had an exponential growth around the world. There are many possible reasons cause this trend increasing rapidly. 
The main problem this paper is going to explore is how is the living conditions in areas with a large number of terrorist attacks? To address this issue, two indexes are introduced, one is global national happiness index (GNH) another one is gross domestic products (GDP). Moreover, the index like family, Life expectancy, freedom, government corruption and dystopia residual will also be introduced to make a compare with the terrorism events.
The difficult is to analyze and present the trends within the dataset as there are too many indexes, and hard to determine which key index matched the terrorism trend most. To solve this problem, provide accurate analysis to which parameters match the terrorism incidents most, these two data set will be used to analyze by following